Rework comparative testing.
